+++ server/libs/domain/src/index.ts
@@ -1,0 +1,2 @@
+export * from './domain.module';
+export * from './user';
