+++ server/libs/infra/src/index.ts
@@ -1,0 +1,2 @@
+export * from './infra.module';
+export * from './repository';
