+++ server/libs/infra/src/repository/index.ts
@@ -1,0 +1,1 @@
+export * from './user.repository';
