+++ server/package.json
@@ -151,7 +151,9 @@
       "^@app/job(|/.*)$": "<rootDir>/libs/job/src/$1",
       "@app/job": "<rootDir>/libs/job/src",
       "^@app/immich-config(|/.*)$": "<rootDir>/libs/immich-config/src/$1",
-      "^@app/storage(|/.*)$": "<rootDir>/libs/storage/src/$1"
+      "^@app/storage(|/.*)$": "<rootDir>/libs/storage/src/$1",
+      "^@app/infra(|/.*)$": "<rootDir>/libs/infra/src/$1",
+      "^@app/domain(|/.*)$": "<rootDir>/libs/domain/src/$1"
     }
   }
 }
