+++ server/tsconfig.json
@@ -45,6 +45,18 @@
       ],
       "@app/storage/*": [
         "libs/storage/src/*"
+      ],
+      "@app/infra": [
+        "libs/infra/src"
+      ],
+      "@app/infra/*": [
+        "libs/infra/src/*"
+      ],
+      "@app/domain": [
+        "libs/domain/src"
+      ],
+      "@app/domain/*": [
+        "libs/domain/src/*"
       ]
     }
   },
